competition/City_CycleGAN/requirements.txt
pillow
numpy
requests
munch
tqdm
scipy
tinyms
mindspore